+++ library/extractor/src/main/java/com/google/android/exoplayer2/extractor/mp4/Sniffer.java
@@ -70,7 +70,7 @@ import java.io.IOException;
    * @throws IOException If an error occurs reading from the input.
    */
   public static boolean sniffFragmented(ExtractorInput input) throws IOException {
-    return sniffInternal(input, true);
+    return sniffInternal(input, /* fragmented= */ true, /* acceptHeic= */ false);
   }
 
   /**
@@ -82,10 +82,24 @@ import java.io.IOException;
    * @throws IOException If an error occurs reading from the input.
    */
   public static boolean sniffUnfragmented(ExtractorInput input) throws IOException {
-    return sniffInternal(input, false);
+    return sniffInternal(input, /* fragmented= */ false, /* acceptHeic= */ false);
   }
 
-  private static boolean sniffInternal(ExtractorInput input, boolean fragmented)
+  /**
+   * Returns whether data peeked from the current position in {@code input} is consistent with the
+   * input being an unfragmented MP4 file.
+   *
+   * @param input The extractor input from which to peek data. The peek position will be modified.
+   * @param acceptHeic Whether {@code true} should be returned for HEIC photos.
+   * @return Whether the input appears to be in the unfragmented MP4 format.
+   * @throws IOException If an error occurs reading from the input.
+   */
+  public static boolean sniffUnfragmented(ExtractorInput input, boolean acceptHeic)
+      throws IOException {
+    return sniffInternal(input, /* fragmented= */ false, acceptHeic);
+  }
+
+  private static boolean sniffInternal(ExtractorInput input, boolean fragmented, boolean acceptHeic)
       throws IOException {
     long inputLength = input.getLength();
     int bytesToSearch = (int) (inputLength == C.LENGTH_UNSET || inputLength > SEARCH_LENGTH
@@ -165,7 +179,7 @@ import java.io.IOException;
           if (i == 1) {
             // This index refers to the minorVersion, not a brand, so skip it.
             buffer.skipBytes(4);
-          } else if (isCompatibleBrand(buffer.readInt())) {
+          } else if (isCompatibleBrand(buffer.readInt(), acceptHeic)) {
             foundGoodFileType = true;
             break;
           }
@@ -185,9 +199,12 @@ import java.io.IOException;
   /**
    * Returns whether {@code brand} is an ftyp atom brand that is compatible with the MP4 extractors.
    */
-  private static boolean isCompatibleBrand(int brand) {
-    // Accept all brands starting '3gp'.
+  private static boolean isCompatibleBrand(int brand, boolean acceptHeic) {
     if (brand >>> 8 == 0x00336770) {
+      // Brand starts with '3gp'.
+      return true;
+    } else if (brand == 0x68656963 && acceptHeic) {
+      // Brand is `heic` and HEIC is supported by the extractor.
       return true;
     }
     for (int compatibleBrand : COMPATIBLE_BRANDS) {